issue18 and also issue13 - reorganized nav from tabs to pages (to facilitate simpler code for automatically jumping from query to source view), and also moved collections dropdown to just above textbox for entering query
